requirements.txt
# FastAPI and web framework
fastapi
uvicorn[standard]
pydantic
pydantic-settings

# Vector Database
chromadb

# Document processing
pypdf  # Fallback PDF library
pymupdf  # PyMuPDF - Best for Chinese text extraction (install as: pip install pymupdf)
pdfplumber  # Good for Chinese and table extraction (optional, but recommended)
markdown
beautifulsoup4
lxml
requests
pyyaml

# RAG and embeddings
langchain
langchain-community
langchain-openai
sentence-transformers
# CPU-only PyTorch (required by sentence-transformers)
# NOTE: Install CPU version manually:
#   Windows/Linux: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
#   macOS: pip install torch torchvision torchaudio
# PyTorch will be installed automatically by sentence-transformers, but for CPU-only builds,
# it's recommended to install it manually first as shown above

# Vector store (optional)
# pinecone-client  # Optional: cloud vector store

# Utilities
python-dotenv
python-multipart
click
httpx  # HTTP client for frontend
pandas  # Data manipulation for Streamlit

# Frontend (Streamlit)
streamlit
plotly
# Note: Streamlit has built-in markdown support, no need for separate editor

# Development dependencies
black
ruff
pytest
pytest-asyncio
pytest-cov

